feat: wip: show privacy banner on all pages
