+++ src/components/privacy-banner/privacy-banner.scss
@@ -1,5 +1,5 @@
-@import "../../../colors";
-@import "../../../frameless";
+@import "../../colors";
+@import "../../frameless";
 
 
 .privacy-banner {
@@ -12,6 +12,7 @@
     align-items: center;
     justify-content: center;
     margin-bottom: 0px;
+    margin-top: 50px;
 
     .privacy-banner-container {
         display: flex;
